refactor(profile): consolidate volunteer and organization profile pages
- Removed separate VolunteerProfilePage and associated components to streamline the profile management process.
- Updated ProfilePage to dynamically render either VolunteerProfileForm or OrganizationProfile based on user role.
- Enhanced error handling with a NotFound component for unsupported roles.
- Deleted unused components related to volunteer profile editing and viewing.
