+++ app/(withLayout)/organisation/opportunities/create/_components/CreateFooter.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import Loading from "@/app/loading";
 import { Button } from "@/components/ui/button";
- 
+import { Loader2 } from "lucide-react";
+
 interface CreateFooterProps {
   onCreate: () => void;
   isLoading: boolean;
@@ -15,14 +15,15 @@ export default function CreateFooter({
   return (
     <div className="flex justify-center sm:justify-end items-center mt-6 px-4 sm:px-0">
       <Button
-        className="bg-blue-600 hover:bg-blue-700 w-full sm:w-auto px-6 py-2 sm:py-2"
+        className="bg-blue-600 cursor-pointer hover:bg-blue-700 w-full sm:w-auto px-6 py-2 sm:py-2"
         onClick={onCreate}
         disabled={isLoading}
       >
         {isLoading ? (
-          <Loading size="medium">
-            <p className="text-gray-600 mt-2">Creating...</p>
-          </Loading>
+          <>
+            <Loader2 className="animate-spin h-4 w-4 mr-2" />
+            Creating Opportunity...
+          </>
         ) : (
           "Create Opportunity"
         )}
