--- app/(withLayout)/volunteer/opportunity/Applied-tasks/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function page() {
-  return <div>hei</div>;
-}
