--- app/(withLayout)/volunteer/opportunity/find/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-export default function page() {
-  return (
-    <div>
-      hei
-      
-    </div>
-  )
-}
-
