+++ components/layout/messages/components/ConversationHeaderOptimized.tsx
@@ -178,7 +178,7 @@ export const ConversationHeaderOptimized: React.FC<ConversationHeaderProps> = ({
                 <>
                   <DropdownMenuItem asChild>
                     <Link
-                      href={`/volunteer/${user._id}/profile`}
+                      href={`/view-profile/volunteer/details/${user._id}`}
                       className="flex items-center cursor-pointer"
                     >
                       <User className="mr-2 h-4 w-4" />
